pyproject.toml
[project]
name = "tushare-a-fundamentals"
version = "0.2.0"
description = "A-share fundamentals downloader using TuShare"
readme = "README.md"
authors = [{ name = "Richard" }]
license = "MIT"
requires-python = ">=3.10,<3.13"
dependencies = [
    "tushare>=1.2.89",
    "pandas>=1.5.0",
    "pyyaml>=6.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
    "pyarrow>=14.0.0",
]

[project.scripts]
funda = "tushare_a_fundamentals.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}
packages = [
    "tushare_a_fundamentals",
    "tushare_a_fundamentals.meta",
    "tushare_a_fundamentals.transforms",
    "tushare_a_fundamentals.writers",
]

[tool.uv]
python-preference = "only-managed"
required-version = ">=0.8.0"
package = true
dev-dependencies = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
    "pytest-cov>=4.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "C90", "I"]
# Ignore rules that conflict with the formatter
ignore = ["E203", "E501"]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.pytest.ini_options]
# Do not force the coverage plugin to avoid failures when pytest-cov is missing
addopts = []
markers = [
    "unit: unit tests",
    "integration: integration tests",
]
testpaths = ["tests"]
